# stop calling memcmp with nullptrs