it doesn't take nullptr according to its declaration in glibc, and calling it in this way causes our sanitizers (ubsan, clang analyze) to fail.

Test Plan:

- command:
```
TEST_TMPDIR=/dev/shm perf record -g ./db_bench -num=1000000 -benchmarks=readwhilewriting -statistics -write_buffer_size=1048576 -target_file_size_base=1048576 -max_bytes_for_level_base=4194304 -compression_type=none -max_background_compactions=8 -max_background_flushes=4
```

- before:
```
readwhilewriting :      20.741 micros/op 48214 ops/sec;    4.4 MB/s (819398 of 1000000 found)
+    9.62%     3.80%  db_bench     db_bench             [.] rocksdb::InternalKeyComparator::Compare
```

- after:
```
readwhilewriting :      19.963 micros/op 50092 ops/sec;    4.4 MB/s (795228 of 1000000 found)
+    9.83%     3.74%  db_bench     db_bench             [.] rocksdb::InternalKeyComparator::Compare
```